Update number idk
Finally fixed the out of gas error on the createDmChat from a turing loop. When a contract tries to update a dynamic array in memory that hasn't got a fixed size, the evm can't push the data and gets struck in a loop. Also added file storage, where most of the heavy lifting will be on the front end.